requirements.txt
streamlit==1.44.1
langchain==0.1.12
langchain-community==0.0.27
langchain-openai==0.0.8
faiss-cpu==1.7.4
pypdf==5.4.0
python-dotenv==1.1.0
openai==1.14.2 